[abi] Rename dir.h to dirent.h
