+++ elks/include/linuxmt/dirent.h
